Add two new core header and refactor to break cycles
