+++ Include/internal/pycore_dtoa.h
@@ -11,53 +11,18 @@ extern "C" {
 #include "pycore_pymath.h"        // _PY_SHORT_FLOAT_REPR
 
 
-typedef uint32_t ULong;
-
-struct
-Bigint {
-    struct Bigint *next;
-    int k, maxwds, sign, wds;
-    ULong x[1];
-};
-
 #if defined(Py_USING_MEMORY_DEBUGGER) || _PY_SHORT_FLOAT_REPR == 0
 
-struct _dtoa_state {
-    int _not_used;
-};
 #define _dtoa_state_INIT(INTERP) \
     {0}
 
-#else  // !Py_USING_MEMORY_DEBUGGER && _PY_SHORT_FLOAT_REPR != 0
+#else
 
-/* The size of the Bigint freelist */
-#define Bigint_Kmax 7
-
-/* The size of the cached powers of 5 array */
-#define Bigint_Pow5size 8
-
-#ifndef PRIVATE_MEM
-#define PRIVATE_MEM 2304
-#endif
-#define Bigint_PREALLOC_SIZE \
-    ((PRIVATE_MEM+sizeof(double)-1)/sizeof(double))
-
-struct _dtoa_state {
-    // p5s is an array of powers of 5 of the form:
-    // 5**(2**(i+2)) for 0 <= i < Bigint_Pow5size
-    struct Bigint *p5s[Bigint_Pow5size];
-    // XXX This should be freed during runtime fini.
-    struct Bigint *freelist[Bigint_Kmax+1];
-    double preallocated[Bigint_PREALLOC_SIZE];
-    double *preallocated_next;
-};
 #define _dtoa_state_INIT(INTERP) \
     { \
         .preallocated_next = (INTERP)->dtoa.preallocated, \
     }
-
-#endif  // !Py_USING_MEMORY_DEBUGGER
-
+#endif
 
 extern double _Py_dg_strtod(const char *str, char **ptr);
 extern char* _Py_dg_dtoa(double d, int mode, int ndigits,
