+++ Include/internal/pycore_function.h
@@ -16,32 +16,10 @@ extern PyObject* _PyFunction_Vectorcall(
     size_t nargsf,
     PyObject *kwnames);
 
-#define FUNC_MAX_WATCHERS 8
 
 #define FUNC_VERSION_UNSET 0
 #define FUNC_VERSION_CLEARED 1
 #define FUNC_VERSION_FIRST_VALID 2
-
-#define FUNC_VERSION_CACHE_SIZE (1<<12)  /* Must be a power of 2 */
-
-struct _func_version_cache_item {
-    PyFunctionObject *func;
-    PyObject *code;
-};
-
-struct _py_func_state {
-#ifdef Py_GIL_DISABLED
-    // Protects next_version
-    PyMutex mutex;
-#endif
-
-    uint32_t next_version;
-    // Borrowed references to function and code objects whose
-    // func_version % FUNC_VERSION_CACHE_SIZE
-    // once was equal to the index in the table.
-    // They are cleared when the function or code object is deallocated.
-    struct _func_version_cache_item func_version_cache[FUNC_VERSION_CACHE_SIZE];
-};
 
 extern PyFunctionObject* _PyFunction_FromConstructor(PyFrameConstructor *constr);
