+++ Include/internal/pycore_import.h
@@ -10,6 +10,7 @@ extern "C" {
 #endif
 
 #include "pycore_lock.h"          // PyMutex
+#include "pycore_runtime_structs.h"  // _import_state
 #include "pycore_hashtable.h"     // _Py_hashtable_t
 
 extern int _PyImport_IsInitialized(PyInterpreterState *);
@@ -30,73 +31,6 @@ extern int _PyImport_FixupBuiltin(
     const char *name,            /* UTF-8 encoded string */
     PyObject *modules
     );
-
-
-struct _import_runtime_state {
-    /* The builtin modules (defined in config.c). */
-    struct _inittab *inittab;
-    /* The most recent value assigned to a PyModuleDef.m_base.m_index.
-       This is incremented each time PyModuleDef_Init() is called,
-       which is just about every time an extension module is imported.
-       See PyInterpreterState.modules_by_index for more info. */
-    Py_ssize_t last_module_index;
-    struct {
-        /* A lock to guard the cache. */
-        PyMutex mutex;
-        /* The actual cache of (filename, name, PyModuleDef) for modules.
-           Only legacy (single-phase init) extension modules are added
-           and only if they support multiple initialization (m_size >= 0)
-           or are imported in the main interpreter.
-           This is initialized lazily in fix_up_extension() in import.c.
-           Modules are added there and looked up in _imp.find_extension(). */
-        _Py_hashtable_t *hashtable;
-    } extensions;
-    /* Package context -- the full module name for package imports */
-    const char * pkgcontext;
-};
-
-struct _import_state {
-    /* cached sys.modules dictionary */
-    PyObject *modules;
-    /* This is the list of module objects for all legacy (single-phase init)
-       extension modules ever loaded in this process (i.e. imported
-       in this interpreter or in any other).  Py_None stands in for
-       modules that haven't actually been imported in this interpreter.
-
-       A module's index (PyModuleDef.m_base.m_index) is used to look up
-       the corresponding module object for this interpreter, if any.
-       (See PyState_FindModule().)  When any extension module
-       is initialized during import, its moduledef gets initialized by
-       PyModuleDef_Init(), and the first time that happens for each
-       PyModuleDef, its index gets set to the current value of
-       a global counter (see _PyRuntimeState.imports.last_module_index).
-       The entry for that index in this interpreter remains unset until
-       the module is actually imported here.  (Py_None is used as
-       a placeholder.)  Note that multi-phase init modules always get
-       an index for which there will never be a module set.
-
-       This is initialized lazily in PyState_AddModule(), which is also
-       where modules get added. */
-    PyObject *modules_by_index;
-    /* importlib module._bootstrap */
-    PyObject *importlib;
-    /* override for config->use_frozen_modules (for tests)
-       (-1: "off", 1: "on", 0: no override) */
-    int override_frozen_modules;
-    int override_multi_interp_extensions_check;
-#ifdef HAVE_DLOPEN
-    int dlopenflags;
-#endif
-    PyObject *import_func;
-    /* The global import lock. */
-    _PyRecursiveMutex lock;
-    /* diagnostic info in PyImport_ImportModuleLevelObject() */
-    struct {
-        int import_level;
-        PyTime_t accumulated;
-        int header;
-    } find_and_load;
-};
 
 #ifdef HAVE_DLOPEN
 #  include <dlfcn.h>              // RTLD_NOW, RTLD_LAZY
