+++ Include/internal/pycore_interp.h
@@ -10,8 +10,7 @@ extern "C" {
 
 #include <stdbool.h>              // bool
 
-#include "pycore_ast_state.h"     // struct ast_state
-#include "pycore_atexit.h"        // struct atexit_state
+#include "pycore_runtime_structs.h"
 #include "pycore_ceval_state.h"   // struct _ceval_state
 #include "pycore_code.h"          // struct callable_cache
 #include "pycore_codecs.h"        // struct codecs_state
@@ -43,26 +42,6 @@ extern "C" {
 #include "pycore_warnings.h"      // struct _warnings_runtime_state
 
 
-struct _Py_long_state {
-    int max_str_digits;
-};
-
-// Support for stop-the-world events. This exists in both the PyRuntime struct
-// for global pauses and in each PyInterpreterState for per-interpreter pauses.
-struct _stoptheworld_state {
-    PyMutex mutex;       // Serializes stop-the-world attempts.
-
-    // NOTE: The below fields are protected by HEAD_LOCK(runtime), not by the
-    // above mutex.
-    bool requested;      // Set when a pause is requested.
-    bool world_stopped;  // Set when the world is stopped.
-    bool is_global;      // Set when contained in PyRuntime struct.
-
-    PyEvent stop_event;  // Set when thread_countdown reaches zero.
-    Py_ssize_t thread_countdown;  // Number of threads that must pause.
-
-    PyThreadState *requester; // Thread that requested the pause (may be NULL).
-};
 
 #ifdef Py_GIL_DISABLED
 // This should be prime but otherwise the choice is arbitrary. A larger value
@@ -70,42 +49,7 @@ struct _stoptheworld_state {
 #  define NUM_WEAKREF_LIST_LOCKS 127
 #endif
 
-/* cross-interpreter data registry */
-
-/* Tracks some rare events per-interpreter, used by the optimizer to turn on/off
-   specific optimizations. */
-typedef struct _rare_events {
-    /* Setting an object's class, obj.__class__ = ... */
-    uint8_t set_class;
-    /* Setting the bases of a class, cls.__bases__ = ... */
-    uint8_t set_bases;
-    /* Setting the PEP 523 frame eval function, _PyInterpreterState_SetFrameEvalFunc() */
-    uint8_t set_eval_frame_func;
-    /* Modifying the builtins,  __builtins__.__dict__[var] = ... */
-    uint8_t builtin_dict;
-    /* Modifying a function, e.g. func.__defaults__ = ..., etc. */
-    uint8_t func_modification;
-} _rare_events;
-
 /* interpreter state */
-
-/* PyInterpreterState holds the global state for one of the runtime's
-   interpreters.  Typically the initial (main) interpreter is the only one.
-
-   The PyInterpreterState typedef is in Include/pytypedefs.h.
-   */
-struct _is {
-
-    /* This struct contains the eval_breaker,
-     * which is by far the hottest field in this struct
-     * and should be placed at the beginning. */
-    struct _ceval_state ceval;
-
-    PyInterpreterState *next;
-
-    int64_t id;
-    Py_ssize_t id_refcount;
-    int requires_idref;
 
 #define _PyInterpreterState_WHENCE_NOTSET -1
 #define _PyInterpreterState_WHENCE_UNKNOWN 0
@@ -115,193 +59,6 @@ struct _is {
 #define _PyInterpreterState_WHENCE_XI 4
 #define _PyInterpreterState_WHENCE_STDLIB 5
 #define _PyInterpreterState_WHENCE_MAX 5
-    long _whence;
-
-    /* Has been initialized to a safe state.
-
-       In order to be effective, this must be set to 0 during or right
-       after allocation. */
-    int _initialized;
-    /* Has been fully initialized via pylifecycle.c. */
-    int _ready;
-    int finalizing;
-
-    uintptr_t last_restart_version;
-    struct pythreads {
-        uint64_t next_unique_id;
-        /* The linked list of threads, newest first. */
-        PyThreadState *head;
-        _PyThreadStateImpl *preallocated;
-        /* The thread currently executing in the __main__ module, if any. */
-        PyThreadState *main;
-        /* Used in Modules/_threadmodule.c. */
-        Py_ssize_t count;
-        /* Support for runtime thread stack size tuning.
-           A value of 0 means using the platform's default stack size
-           or the size specified by the THREAD_STACK_SIZE macro. */
-        /* Used in Python/thread.c. */
-        size_t stacksize;
-    } threads;
-
-    /* Reference to the _PyRuntime global variable. This field exists
-       to not have to pass runtime in addition to tstate to a function.
-       Get runtime from tstate: tstate->interp->runtime. */
-    struct pyruntimestate *runtime;
-
-    /* Set by Py_EndInterpreter().
-
-       Use _PyInterpreterState_GetFinalizing()
-       and _PyInterpreterState_SetFinalizing()
-       to access it, don't access it directly. */
-    PyThreadState* _finalizing;
-    /* The ID of the OS thread in which we are finalizing. */
-    unsigned long _finalizing_id;
-
-    struct _gc_runtime_state gc;
-
-    /* The following fields are here to avoid allocation during init.
-       The data is exposed through PyInterpreterState pointer fields.
-       These fields should not be accessed directly outside of init.
-
-       All other PyInterpreterState pointer fields are populated when
-       needed and default to NULL.
-
-       For now there are some exceptions to that rule, which require
-       allocation during init.  These will be addressed on a case-by-case
-       basis.  Also see _PyRuntimeState regarding the various mutex fields.
-       */
-
-    // Dictionary of the sys module
-    PyObject *sysdict;
-
-    // Dictionary of the builtins module
-    PyObject *builtins;
-
-    struct _import_state imports;
-
-    /* The per-interpreter GIL, which might not be used. */
-    struct _gil_runtime_state _gil;
-
-     /* ---------- IMPORTANT ---------------------------
-     The fields above this line are declared as early as
-     possible to facilitate out-of-process observability
-     tools. */
-
-    struct codecs_state codecs;
-
-    PyConfig config;
-    unsigned long feature_flags;
-
-    PyObject *dict;  /* Stores per-interpreter state */
-
-    PyObject *sysdict_copy;
-    PyObject *builtins_copy;
-    // Initialized to _PyEval_EvalFrameDefault().
-    _PyFrameEvalFunction eval_frame;
-
-    PyFunction_WatchCallback func_watchers[FUNC_MAX_WATCHERS];
-    // One bit is set for each non-NULL entry in func_watchers
-    uint8_t active_func_watchers;
-
-    Py_ssize_t co_extra_user_count;
-    freefunc co_extra_freefuncs[MAX_CO_EXTRA_USERS];
-
-    /* cross-interpreter data and utils */
-    _PyXI_state_t xi;
-
-#ifdef HAVE_FORK
-    PyObject *before_forkers;
-    PyObject *after_forkers_parent;
-    PyObject *after_forkers_child;
-#endif
-
-    struct _warnings_runtime_state warnings;
-    struct atexit_state atexit;
-    struct _stoptheworld_state stoptheworld;
-    struct _qsbr_shared qsbr;
-
-#if defined(Py_GIL_DISABLED)
-    struct _mimalloc_interp_state mimalloc;
-    struct _brc_state brc;  // biased reference counting state
-    struct _Py_unique_id_pool unique_ids;  // object ids for per-thread refcounts
-    PyMutex weakref_locks[NUM_WEAKREF_LIST_LOCKS];
-    _PyIndexPool tlbc_indices;
-#endif
-    // Per-interpreter list of tasks, any lingering tasks from thread
-    // states gets added here and removed from the corresponding
-    // thread state's list.
-    struct llist_node asyncio_tasks_head;
-    // `asyncio_tasks_lock` is used when tasks are moved
-    // from thread's list to interpreter's list.
-    PyMutex asyncio_tasks_lock;
-
-    // Per-interpreter state for the obmalloc allocator.  For the main
-    // interpreter and for all interpreters that don't have their
-    // own obmalloc state, this points to the static structure in
-    // obmalloc.c obmalloc_state_main.  For other interpreters, it is
-    // heap allocated by _PyMem_init_obmalloc() and freed when the
-    // interpreter structure is freed.  In the case of a heap allocated
-    // obmalloc state, it is not safe to hold on to or use memory after
-    // the interpreter is freed. The obmalloc state corresponding to
-    // that allocated memory is gone.  See free_obmalloc_arenas() for
-    // more comments.
-    struct _obmalloc_state *obmalloc;
-
-    PyObject *audit_hooks;
-    PyType_WatchCallback type_watchers[TYPE_MAX_WATCHERS];
-    PyCode_WatchCallback code_watchers[CODE_MAX_WATCHERS];
-    PyContext_WatchCallback context_watchers[CONTEXT_MAX_WATCHERS];
-    // One bit is set for each non-NULL entry in code_watchers
-    uint8_t active_code_watchers;
-    uint8_t active_context_watchers;
-
-    struct _py_object_state object_state;
-    struct _Py_unicode_state unicode;
-    struct _Py_long_state long_state;
-    struct _dtoa_state dtoa;
-    struct _py_func_state func_state;
-    struct _py_code_state code_state;
-
-    struct _Py_dict_state dict_state;
-    struct _Py_exc_state exc_state;
-    struct _Py_mem_interp_free_queue mem_free_queue;
-
-    struct ast_state ast;
-    struct types_state types;
-    struct callable_cache callable_cache;
-    bool jit;
-    _PyExecutorObject *executor_list_head;
-    size_t trace_run_counter;
-    _rare_events rare_events;
-    PyDict_WatchCallback builtins_dict_watcher;
-
-    _Py_GlobalMonitors monitors;
-    bool sys_profile_initialized;
-    bool sys_trace_initialized;
-    Py_ssize_t sys_profiling_threads; /* Count of threads with c_profilefunc set */
-    Py_ssize_t sys_tracing_threads; /* Count of threads with c_tracefunc set */
-    PyObject *monitoring_callables[PY_MONITORING_TOOL_IDS][_PY_MONITORING_EVENTS];
-    PyObject *monitoring_tool_names[PY_MONITORING_TOOL_IDS];
-    uintptr_t monitoring_tool_versions[PY_MONITORING_TOOL_IDS];
-
-    struct _Py_interp_cached_objects cached_objects;
-    struct _Py_interp_static_objects static_objects;
-
-    Py_ssize_t _interactive_src_count;
-
-    /* the initial PyInterpreterState.threads.head */
-    _PyThreadStateImpl _initial_thread;
-    // _initial_thread should be the last field of PyInterpreterState.
-    // See https://github.com/python/cpython/issues/127117.
-
-#if !defined(Py_GIL_DISABLED) && defined(Py_STACKREF_DEBUG)
-    uint64_t next_stackref;
-    _Py_hashtable_t *open_stackrefs_table;
-#  ifdef Py_STACKREF_CLOSE_DEBUG
-    _Py_hashtable_t *closed_stackrefs_table;
-#  endif
-#endif
-};
 
 
 /* other API */
