+++ Include/internal/pycore_pyhash.h
@@ -71,19 +71,6 @@ extern int _Py_HashSecret_Initialized;
 #endif
 
 
-struct pyhash_runtime_state {
-    struct {
-#ifndef MS_WINDOWS
-        int fd;
-        dev_t st_dev;
-        ino_t st_ino;
-#else
-    // This is a placeholder so the struct isn't empty on Windows.
-    int _not_used;
-#endif
-    } urandom_cache;
-};
-
 #ifndef MS_WINDOWS
 # define _py_urandom_cache_INIT \
     { \
