+++ Python/dtoa.c
@@ -122,6 +122,8 @@
 #include "pycore_pystate.h"       // _PyInterpreterState_GET()
 #include <stdlib.h>               // exit()
 
+typedef uint32_t ULong;
+
 /* if _PY_SHORT_FLOAT_REPR == 0, then don't even try to compile
    the following code */
 #if _PY_SHORT_FLOAT_REPR == 1
